streamlit_app/requirements.txt
catboost==1.2.5
imblearn==0.0
joblib==1.3.2
keras==3.3.3
matplotlib==3.8.0
numpy==1.26.3
pandas==2.1.4
plotly==5.20.0
prophet==1.1.5
pyarrow==15.0.2
pytorch-tabnet==4.1.0
PyYAML==6.0.1
scikit-learn==1.4.1.post1
seaborn==0.13.0
statsforecast==1.7.3
statsmodels==0.14.0
streamlit==1.35.0
tensorflow==2.16.1
torch==2.3.1
xgboost==2.0.3